// src/frontend/src/extensions/sim_metadata.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SimulationDataExtensionMetadata {
  /**
   * Name of the simulation.
   */
  name: string;
  /**
   * Date and time of the simulation.
   */
  date: string;
  /**
   * Name of the FEA software used.
   */
  fea_software: string;
  /**
   * Version of the FEA software.
   */
  fea_software_version: string;
  /**
   * List of analysis steps.
   */
  steps: StepObject[];
  [k: string]: unknown;
}
export interface StepObject {
  /**
   * Type of analysis for this step.
   */
  analysis_type: "eigenvalue" | "implicit static" | "implicit dynamic" | "explicit";
  /**
   * List of field data objects for this step.
   */
  fields: FieldObject[];
  [k: string]: unknown;
}
export interface FieldObject {
  /**
   * Name of the result field, e.g., displacement, stress.
   */
  name: string;
  /**
   * Type of the field data, e.g., 'vector', 'scalar'.
   */
  type: string;
  data: DataReference;
  [k: string]: unknown;
}
export interface DataReference {
  /**
   * Index of the glTF bufferView containing the binary data.
   */
  bufferView: number;
  /**
   * Offset into the bufferView in bytes.
   */
  byteOffset?: number;
  [k: string]: unknown;
}